WIP for the new structure of the entire codebase
