+++ src/app/blog/not-found.tsx
@@ -1,0 +1,3 @@
+export default function NotFound() {
+  return <div>Not Found</div>;
+}
