+++ src/components/sections/references-section.tsx
@@ -3,9 +3,9 @@
 import React from 'react';
 import { Flex } from '@radix-ui/themes';
 
-import ProjectCard from '../ProjectCard';
-import SectionHeader from '../SectionHeader';
-import { useLanguage } from '@/i18n/LanguageProvider';
+import ProjectCard from '../ui/project-card';
+import SectionHeader from '../layout/section-header';
+import { useLanguage } from '@/components/providers/language-providers';
 
 export default function ReferencesStrategy() {
   const { messages } = useLanguage();
