+++ src/lib/hooks/use-theme.ts
@@ -1,0 +1,10 @@
+import { ThemeContext } from '@/components/providers/theme-provider';
+import { useContext } from 'react';
+
+export function useTheme() {
+  const context = useContext(ThemeContext);
+  if (context === undefined) {
+    throw new Error('useTheme must be used within a ThemeProvider');
+  }
+  return context;
+}
