+++ src/lib/services/firebase/config.ts
@@ -1,7 +1,4 @@
-import { initializeApp } from 'firebase/app';
-import { getStorage } from 'firebase/storage';
-
-const firebaseConfig = {
+export const firebaseConfig = {
   apiKey: 'AIzaSyAvGhqOoUyZ26NX1IgIj9Tfhzx82KSaiEk',
   authDomain: 'portfolio-c336c.firebaseapp.com',
   projectId: 'portfolio-c336c',
@@ -10,11 +7,3 @@ const firebaseConfig = {
   appId: '1:527681504318:web:a561cf6d8187c9e2307d6f',
   measurementId: 'G-N4V8S4JMN2',
 };
-
-// Initialize Firebase
-const app = initializeApp(firebaseConfig);
-
-// Initialize Firebase Storage
-export const storage = getStorage(app);
-
-export default app;
