+++ src/lib/types/theme.ts
@@ -1,0 +1,6 @@
+export type ThemeMode = 'dark' | 'light';
+
+export type ThemeContextType = {
+  theme: ThemeMode;
+  toggleTheme: () => void;
+};
